Lots of Alpha 4 Progress
The final trial planned for the 2.0 release is going to be in A4, although that has not been started yet. For now, huge reworks to the custom food and storage systems, and lots of improvements to the current trials in terms of look and feel. Not textures, but plants and biomes and stuff! Next alpha will include the new trial though! At least, the start of it.